Move mod.rs stuff into backend.rs
This is the beginning of the backend change.
