+++ src/backend/mod.rs
@@ -17,164 +17,13 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#[allow(clippy::module_inception)]
+pub mod backend;
 pub mod cryfs;
 pub mod gocryptfs;
 
-use crate::vault::VaultConfig;
-use gettextrs::gettext;
-use serde::{Deserialize, Serialize};
-use std::string::String;
-use strum_macros::EnumIter;
-
-pub const AVAILABLE_BACKENDS: &'static [&'static str] = &["gocryptfs", "cryfs"];
-
-quick_error! {
-    #[derive(Debug)]
-    pub enum BackendError {
-        ToUser(e: String) {
-            display("{}", e)
-        }
-        Generic {
-            from(std::io::Error)
-        }
-    }
-}
-
-#[derive(
-    Debug,
-    EnumIter,
-    strum_macros::Display,
-    Serialize,
-    Deserialize,
-    strum_macros::EnumString,
-    Copy,
-    Clone,
-)]
-pub enum Backend {
-    Cryfs,
-    Gocryptfs,
-}
-
-impl Backend {
-    pub fn is_available(&self) -> Result<bool, BackendError> {
-        match &self {
-            Backend::Cryfs => cryfs::is_available(),
-            Backend::Gocryptfs => gocryptfs::is_available(),
-        }
-    }
-
-    pub fn init(vault_config: &VaultConfig, password: String) -> Result<(), BackendError> {
-        let encrypted_data_directory = &vault_config.encrypted_data_directory;
-        let mount_directory = &vault_config.mount_directory;
-
-        match create_edd_if_not_exists(&encrypted_data_directory) {
-            Ok(_) => {}
-            Err(e) => {
-                return Err(e);
-            }
-        }
-
-        match create_md_if_not_exists(&mount_directory) {
-            Ok(_) => {}
-            Err(e) => {
-                return Err(e);
-            }
-        }
-
-        match vault_config.backend {
-            Backend::Cryfs => cryfs::init(vault_config, password),
-            Backend::Gocryptfs => gocryptfs::init(vault_config, password),
-        }
-    }
-
-    pub fn open(vault_config: &VaultConfig, password: String) -> Result<(), BackendError> {
-        match vault_config.backend {
-            Backend::Cryfs => cryfs::open(vault_config, password),
-            Backend::Gocryptfs => gocryptfs::open(vault_config, password),
-        }
-    }
-
-    pub fn close(vault_config: &VaultConfig) -> Result<(), BackendError> {
-        match vault_config.backend {
-            Backend::Cryfs => cryfs::close(vault_config),
-            Backend::Gocryptfs => gocryptfs::close(vault_config),
-        }
-    }
-}
-
-pub fn get_ui_string_from_backend(backend: &Backend) -> String {
-    match backend {
-        Backend::Cryfs => String::from("CryFS"),
-        Backend::Gocryptfs => String::from("gocryptfs"),
-    }
-}
-
-pub fn get_backend_from_ui_string(backend: &String) -> Option<Backend> {
-    if backend == "CryFS" {
-        Some(Backend::Cryfs)
-    } else if backend == "gocryptfs" {
-        Some(Backend::Gocryptfs)
-    } else {
-        None
-    }
-}
-
-fn create_edd_if_not_exists(encrypted_data_directory: &String) -> Result<(), BackendError> {
-    match std::fs::create_dir_all(encrypted_data_directory) {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            log::debug!("Failed to create encrypted data directory: {}", e);
-
-            match e.kind() {
-                std::io::ErrorKind::PermissionDenied => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory: Permission denied.",
-                ))),
-                std::io::ErrorKind::ConnectionRefused => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory: Connection refused.",
-                ))),
-                std::io::ErrorKind::ConnectionReset => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory: Connection reset.",
-                ))),
-                std::io::ErrorKind::ConnectionAborted => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory: Connection aborted.",
-                ))),
-                std::io::ErrorKind::NotConnected => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory: Not connected.",
-                ))),
-                _ => Err(BackendError::ToUser(gettext(
-                    "Failed to create encrypted data directory.",
-                ))),
-            }
-        }
-    }
-}
-
-fn create_md_if_not_exists(mount_directory: &String) -> Result<(), BackendError> {
-    match std::fs::create_dir_all(mount_directory) {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            log::debug!("Failed to create encrypted data directory: {}", e);
-
-            match e.kind() {
-                std::io::ErrorKind::PermissionDenied => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory: Permission denied.",
-                ))),
-                std::io::ErrorKind::ConnectionRefused => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory: Connection refused.",
-                ))),
-                std::io::ErrorKind::ConnectionReset => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory: Connection reset.",
-                ))),
-                std::io::ErrorKind::ConnectionAborted => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory: Connection aborted.",
-                ))),
-                std::io::ErrorKind::NotConnected => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory: Not connected.",
-                ))),
-                _ => Err(BackendError::ToUser(gettext(
-                    "Failed to create mount directory.",
-                ))),
-            }
-        }
-    }
-}
+pub use backend::get_backend_from_ui_string;
+pub use backend::get_ui_string_from_backend;
+pub use backend::Backend;
+pub use backend::BackendError;
+pub use backend::AVAILABLE_BACKENDS;
